backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
executing==1.2.0
inflection==0.5.1
ipython==8.12.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
wcwidth==0.2.6
alabaster==0.7.13
aniso8601==9.0.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
astroid==2.14.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.2
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitstring==3.1.9
black==23.3.0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cookiecutter==2.1.1
coverage==7.1.0
cpufreq==0.3.5
crispy-bootstrap5==0.7
cryptography==38.0.3
cupshelpers==1.0
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
django-allauth==0.52.0
django-braces==1.15.0
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-coverage-plugin==3.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.14.0
django-stubs-ext==0.7.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
docutils==0.19
drf-spectacular==0.25.1
duplicity==0.8.21
entrypoints==0.4
evdev==1.6.0
exceptiongroup==1.1.0
executing==1.0.0
factory-boy==3.2.1
Faker==16.6.1
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.8.0
flake8==6.0.0
flake8-isort==6.0.0
Flask==2.2.3
Flask-RESTful==0.3.9
fonttools==4.37.1
future==0.18.2
ghalton==0.6.1
hiredis==2.1.1
httplib2==0.20.2
hyperlink==21.0.0
identify==2.5.17
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.11
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
meson==0.61.2
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
mypy==0.982
mypy-extensions==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
nodeenv==1.7.0
notebook==6.4.12
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.4.4
pandocfilters==1.5.0
paramiko==2.9.3
parsecpy==1.0.6
parso==0.8.3
pathspec==0.11.0
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==3.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pyflakes==3.0.1
Pygments==2.13.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.16.1
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.1
pytest-django==4.5.2
pytest-sugar==0.9.6
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-Levenshtein==0.12.2
python-slugify==7.0.0
python-xlib==0.32
python3-openid==3.2.0
pytz==2022.7.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
redis==4.4.2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.5.0
systemd-python==234
termcolor==2.2.0
terminado==0.15.0
terminedia==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.3.0
Twisted==22.10.0
txaio==22.2.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.4
types-requests==2.28.11.10
types-urllib3==1.26.25.4
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ulauncher==5.15.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.6
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xarray==2022.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.5.0
